updated models

+++ suntransfers/models.py
@@ -1,6 +1,5 @@
-from sqlalchemy import Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import create_engine, Column, Integer, String, Boolean, Sequence, UniqueConstraint
+from sqlalchemy import create_engine, Column, Integer, String, Boolean, Sequence, UniqueConstraint,Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
@@ -20,22 +19,22 @@ class Route(Base):
     code = Column(String(10))
     Route_dest = Column(String(255))
     to_alternateId = Column(Integer)
-    pax_1 = Column(Integer)
-    pax_2 = Column(Integer)
-    pax_3 = Column(Integer)
-    pax_4 = Column(Integer)
-    pax_5 = Column(Integer)
-    pax_6 = Column(Integer)
-    pax_7 = Column(Integer)
-    pax_8 = Column(Integer)
-    pax_9 = Column(Integer)
-    pax_10 = Column(Integer)
-    pax_11 = Column(Integer)
-    pax_12 = Column(Integer)
-    pax_13 = Column(Integer)
-    pax_14 = Column(Integer)
-    pax_15 = Column(Integer)
-    pax_16 = Column(Integer)
+    pax_1 = Column(Float)
+    pax_2 = Column(Float)
+    pax_3 = Column(Float)
+    pax_4 = Column(Float)
+    pax_5 = Column(Float)
+    pax_6 = Column(Float)
+    pax_7 = Column(Float)
+    pax_8 = Column(Float)
+    pax_9 = Column(Float)
+    pax_10 = Column(Float)
+    pax_11 = Column(Float)
+    pax_12 = Column(Float)
+    pax_13 = Column(Float)
+    pax_14 = Column(Float)
+    pax_15 = Column(Float)
+    pax_16 = Column(Float)
 
     __table_args__ = (UniqueConstraint('key', name='uq_key'),)
